cleans up styles and adds missing propTypes
