+++ app/javascript/src/application/uiElements/Button.jsx
@@ -2,19 +2,20 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { withStyles } from '@material-ui/core/styles';
 import Button from '@material-ui/core/Button';
-
-const styles = theme => ({
-  button: {
-    margin: theme.spacing.unit,
-  }
-});
+import { buttonStyles } from './styles';
 
 function ContainedButton(props) {
   const { classes } = props;
-
   return (
     <Button
-      variant="contained" className={classes.button} size={props.size} color={props.color} onClick={props.onClick} disabled={props.disabled} type={props.type}>
+      variant="contained"
+      className={classes.button}
+      size={props.size}
+      color={props.color}
+      onClick={props.onClick}
+      disabled={props.disabled}
+      type={props.type}
+    >
       {props.children}
     </Button >
   );
@@ -24,4 +25,4 @@ ContainedButton.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(ContainedButton);
+export default withStyles(buttonStyles)(ContainedButton);
